+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - On the Profile screen, open a sheet to display the full bio. 
+- Fixed an issue where "Share logs" wasn't working.
 - On the Discover tab, center the category buttons.
 
 ## [0.1.12] - 2024-05-07Z