+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Prefetch reply, like and repost counts while displaying the feeds.
 - Fixed an issue where "Share logs" wasn't working.
 - Discover tab: Center the category buttons.
 - Discover tab: Remove placeholder categories and people.