+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- On the Profile screen, open a sheet to display the full bio. 
+- On the Profile screen, open a sheet to display the full bio.
+- Add the author's name to profile cards on the Discover tab and search results. 
 - Fixed an issue where "Share logs" wasn't working.
 - Discover tab: Center the category buttons.
 - Discover tab: Remove placeholder categories and people.